fix(admin): fix runtime bugs - entity type mismatch and security auth
1. StageLogStatsRead: change avgMs/p95Ms from Double to Long and
   successCount/failedCount from Integer to Long to match MySQL bigint
   columns (ddl-auto=validate was failing on startup).
2. SecurityConfig: set SecurityContext Authentication after token
   validation, otherwise AuthorizationFilter returns 403 even with
   valid token.
3. Update StageDurationDTO and MetricsAdminServiceTest to match.

+++ agent-admin-backend/src/main/java/com/meichen/admin/dto/StageDurationDTO.java
@@ -2,9 +2,9 @@ package com.meichen.admin.dto;
 
 public record StageDurationDTO(
     String stageName,
-    Double avgMs,
-    Double p95Ms,
+    Long avgMs,
+    Long p95Ms,
     Long maxMs,
-    Integer successCount,
-    Integer failedCount
+    Long successCount,
+    Long failedCount
 ) {}
